Use size_t for the sizes instead of int

+++ src/DutagaciDialog.cpp
@@ -84,7 +84,7 @@ void DutagaciDialog::on_addButton_clicked()
             if(fileName.right(3) == "txt")
             {
                 std::vector<QString> objFiles = GetFileListFromTXT(fileName);
-                for( int j = 0; j < objFiles.size(); j++ )
+                for( size_t j = 0; j < objFiles.size(); j++ )
                 {
                     mUi->filesListWidget->addItem(objFiles.at(j));
                 }
